1_spark_datapipeline/requirements.txt
pydicom==2.3.0
pillow==9.1.0
numpy==1.21.5